refactor: GameTable 삭제, Game으로 편입
