+++ src/main/java/blackjack/domain/Game.java
@@ -4,8 +4,7 @@ import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Participant;
 import blackjack.domain.gamer.Player;
 import blackjack.domain.gamer.Players;
-import blackjack.domain.gametable.GameTable;
-import blackjack.domain.utils.RandomCardDeck;
+import blackjack.domain.utils.CardDeck;
 import blackjack.dto.ProcessDto;
 import blackjack.dto.ResultDto;
 import java.util.ArrayList;
@@ -14,15 +13,24 @@ import java.util.List;
 public class Game {
     public static final String DELIMITER = ",";
 
-    private final GameTable gameTable;
+
     private final Players players;
     private final Participant dealer;
+    private final CardDeck cardDeck;
 
-    public Game(final List<String> participantsInfo) {
+    public Game(final List<String> participantsInfo, CardDeck cardDeck) {
         List<Player> playersValue = getPlayerList(participantsInfo);
         this.dealer = new Dealer();
         this.players = new Players(playersValue);
-        this.gameTable = new GameTable(dealer, players, new RandomCardDeck());
+        this.cardDeck = cardDeck;
+        initGame();
+    }
+
+    public Game(final Participant dealer, Players players, CardDeck cardDeck) {
+        this.dealer = dealer;
+        this.players = players;
+        this.cardDeck = cardDeck;
+        initGame();
     }
 
     private List<Player> getPlayerList(final List<String> participantsInfo) {
@@ -36,6 +44,13 @@ public class Game {
         return playersValue;
     }
 
+    private void initGame() {
+        dealer.takeCard(cardDeck.pop());
+        dealer.takeCard(cardDeck.pop());
+
+        players.takeTwoCards(cardDeck);
+    }
+
     public ProcessDto getProcessDto() {
         return new ProcessDto(players, dealer);
     }
@@ -45,12 +60,16 @@ public class Game {
     }
 
     public Player turnForPlayer(final Player player) {
-        gameTable.giveCard(player);
+        if (player.isAbleToTake()) {
+            player.takeCard(cardDeck.pop());
+        }
         return player;
     }
 
     public void turnForDealer() {
-        gameTable.giveCard(dealer);
+        if (dealer.isAbleToTake()) {
+            dealer.takeCard(cardDeck.pop());
+        }
     }
 
     public ResultDto getResultDto() {
