+++ src/main/java/blackjack/dto/PlayerResult.java
@@ -1,8 +1,8 @@
 package blackjack.dto;
 
+import blackjack.domain.Outcome;
 import blackjack.domain.gamer.BettingMoney;
 import blackjack.domain.gamer.Player;
-import blackjack.domain.gametable.Outcome;
 
 public class PlayerResult {
     private final String name;
