+++ src/test/java/blackjack/domain/OutcomeTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gametable;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
